requirements.txt
dj-database-url==0.4.1
dj-static==0.0.6
django==1.10.2
django-cors-headers==1.2.2
django-filter==0.15.2
djangorestframework==3.4.7
gunicorn==19.6.0
Markdown==2.6.7
psycopg2==2.6.2
pytz==2016.7
static3==0.7.0
Pillow==3.4.0
